정상적인 플레이가 가능하도록 경과시간 구하기 수정
문제점: 미리 로드한 프레임을 정상적으로 다 사용하지 않는 경향이 있음

+++ Video.h
@@ -30,7 +30,7 @@ public:
     return m_fp;
   }
 
-  int getCurrentFrameIDByElapsed(uint64_t elapsed) const;
+  int getCurrentFrameIDByElapsed(double elapsed) const;
 
   bool readVideoFromFile(const char* path);
